# docs/swagger.yaml
basePath: /api/v1
definitions:
  forms.EditFood:
    properties:
      image:
        type: string
      menu_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  forms.Food:
    properties:
      image:
        type: string
      menu_id:
        type: string
      name:
        minLength: 1
        type: string
      price:
        minimum: 0
        type: number
    required:
    - menu_id
    - name
    - price
    type: object
...
  models.Food:
    properties:
      date_created:
        type: string
      date_updated:
        type: string
      id:
        type: string
      image:
        type: string
      menu_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  response.SuccessMessage:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: fetched successfully
        type: string
      status:
        example: success
        type: string
    type: object
  se.ServiceError:
    properties:
      description:
        type: string
      error:
        type: string
      time:
        type: string
      type:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: bobdence@gmail.com
    name: Confidence James
    url: http://github.com/jamesconfy
  description: Server for {{your server link}}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Project
  version: "1.0"
paths:
...
  /foods:
    get:
      consumes:
      - application/json
      description: Provide details to get all food
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessMessage'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Food'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/se.ServiceError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/se.ServiceError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/se.ServiceError'
      summary: Get all Food route
      tags:
      - Food
    post:
      consumes:
      - application/json
      description: Provide details to add food
      parameters:
      - description: Food details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forms.Food'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessMessage'
            - properties:
                data:
                  $ref: '#/definitions/models.Food'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/se.ServiceError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/se.ServiceError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/se.ServiceError'
      security:
      - ApiKeyAuth: []
      summary: Add Food route
      tags:
      - Food
  /foods/{foodId}:
    delete:
      consumes:
      - application/json
      description: Provide details to delete food
      parameters:
      - description: Food Id
        format: uuid
        in: path
        name: foodId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: food deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/se.ServiceError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/se.ServiceError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/se.ServiceError'
      security:
      - ApiKeyAuth: []
      summary: Delete Food route
      tags:
      - Food
    get:
      consumes:
      - application/json
      description: Provide details to get food
      parameters:
      - description: Food Id
        format: uuid
        in: path
        name: foodId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessMessage'
            - properties:
                data:
                  $ref: '#/definitions/models.Food'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/se.ServiceError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/se.ServiceError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/se.ServiceError'
      summary: Get Food route
      tags:
      - Food
    patch:
      consumes:
      - application/json
      description: Provide details to edit food
      parameters:
      - description: Food Id
        format: uuid
        in: path
        name: foodId
        required: true
        type: string
      - description: Food details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forms.EditFood'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessMessage'
            - properties:
                data:
                  $ref: '#/definitions/models.Food'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/se.ServiceError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/se.ServiceError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/se.ServiceError'
      security:
      - ApiKeyAuth: []
      summary: Edit Food route
      tags:
      - Food
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorisation
    type: apiKey
swagger: "2.0"